refactor: rename state management files for consistency and clarity
